requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-chroma>=0.1.0
langgraph>=0.1.0
openai>=1.0.0
tiktoken>=0.5.0
chromadb>=0.4.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
